requirements.txt
numpy
pillow
imageio
imageio-ffmpeg
gradio
